Rewrite client interface for login server (part 6)
Moved login clif code to lclif.c/lclif.h
