+++ vcproj-10/login-server.vcxproj
@@ -156,6 +156,7 @@
     <ClInclude Include="..\src\login\account.h" />
     <ClInclude Include="..\src\login\HPMlogin.h" />
     <ClInclude Include="..\src\login\ipban.h" />
+    <ClInclude Include="..\src\login\lclif.h" />
     <ClInclude Include="..\src\login\login.h" />
     <ClInclude Include="..\src\login\loginlog.h" />
     <ClInclude Include="..\src\common\cbasetypes.h" />
@@ -192,6 +193,7 @@
     <ClCompile Include="..\src\login\account_sql.c" />
     <ClCompile Include="..\src\login\HPMlogin.c" />
     <ClCompile Include="..\src\login\ipban_sql.c" />
+    <ClCompile Include="..\src\login\lclif.c" />
     <ClCompile Include="..\src\login\login.c" />
     <ClCompile Include="..\src\login\loginlog_sql.c" />
     <ClCompile Include="..\src\common\conf.c" />
