+++ vcproj-12/login-server.vcxproj.filters
@@ -94,6 +94,9 @@
     <ClCompile Include="..\src\common\des.c">
       <Filter>common</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\login\lclif.c">
+      <Filter>login</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\src\login\account.h">
@@ -236,6 +239,9 @@
     </ClInclude>
     <ClInclude Include="..\src\config\classes\general.h">
       <Filter>config</Filter>
+    </ClInclude>
+    <ClInclude Include="..\src\login\lclif.h">
+      <Filter>login</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
