+++ vcproj-14/login-server.vcxproj
@@ -158,6 +158,7 @@
     <ClInclude Include="..\src\login\account.h" />
     <ClInclude Include="..\src\login\HPMlogin.h" />
     <ClInclude Include="..\src\login\ipban.h" />
+    <ClInclude Include="..\src\login\lclif.h" />
     <ClInclude Include="..\src\login\login.h" />
     <ClInclude Include="..\src\login\loginlog.h" />
     <ClInclude Include="..\src\common\cbasetypes.h" />
@@ -194,6 +195,7 @@
     <ClCompile Include="..\src\login\account_sql.c" />
     <ClCompile Include="..\src\login\HPMlogin.c" />
     <ClCompile Include="..\src\login\ipban_sql.c" />
+    <ClCompile Include="..\src\login\lclif.c" />
     <ClCompile Include="..\src\login\login.c" />
     <ClCompile Include="..\src\login\loginlog_sql.c" />
     <ClCompile Include="..\src\common\conf.c" />
